+++ src/Bumblebee/Resources/Collection.js
@@ -6,6 +6,7 @@ const ResourceAbstract = require('./ResourceAbstract')
  * Collection class
  *
  * @class Collection
+ * @extends ResourceAbstract
  */
 class Collection extends ResourceAbstract {
 }